+++ CHANGES.md
@@ -54,6 +54,7 @@
 * Remove FreshOnTV (TvT) torrent provider
 * Remove Strike torrent provider
 * Change increase the scope and number of non release group text that is identified and removed
+* Add a general config setting to allow adding incomplete show data
 
 [develop changelog]
 Enable Alpha Ratio again now that the secure login page over https is fixed